api/requirements.txt
Flask
pymongo
requests
redis
pyjwt
flask-cors